refactor: renamed node and network spawner files
