+++ ant-node/examples/spawn_local_network.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use ant_node::spawn::network::NetworkSpawner;
+use ant_node::spawn::network_spawner::NetworkSpawner;
 use std::time::Duration;
 use tokio::time::sleep;
 use tracing_subscriber::layer::SubscriberExt;
