+++ ant-node/src/spawn/node_spawner.rs
